+++ build.gradle
@@ -134,6 +134,10 @@ dependencies {
 
 	// Quartz
 	implementation 'org.springframework.boot:spring-boot-starter-quartz'
+
+	// feign
+	implementation platform("org.springframework.cloud:spring-cloud-dependencies:2025.0.0")
+	implementation "org.springframework.cloud:spring-cloud-starter-openfeign"
 }
 
 tasks.named('test') {
