+++ src/main/java/com/grow/member_service/MemberServiceApplication.java
@@ -2,8 +2,10 @@ package com.grow.member_service;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.openfeign.EnableFeignClients;
 import org.springframework.scheduling.annotation.EnableScheduling;
 
+@EnableFeignClients
 @EnableScheduling
 @SpringBootApplication
 public class MemberServiceApplication {
